keep_coords.dtype -> np.int_
Fixes #26

- The seed coordinates for structures that are kept for volume visualization weren't kept as an integer dtype. This is fixed during volume during the volume visualization when creating an array from the coords.

+++ library/volume_visualization.py
@@ -76,7 +76,7 @@ def vol_plot(volume, g, build_original, build_smoothed, status_updater=None,
         keep_coords.append(g.vs[c[0]]['v_coords'])     
     
     # For visualization purposes, filter the deleted segments' voxels
-    keep_coords = np.array(keep_coords)
+    keep_coords = np.array(keep_coords, dtype=np.int_)
     volume = VolProc.filter_volume(volume, keep_coords)
     
     if verbose:
